PWA-2510 fixed and optimized imports and exports of types.
