+++ libraries/engage/cart/components/CartItems/CartItemGroupReservationLabel.jsx
@@ -2,7 +2,7 @@
 import * as React from 'react';
 import { i18n } from '@shopgate/engage/core';
 import { LocationIcon } from '@shopgate/engage/components';
-import { type Location } from '@shopgate/engage/locations/locations.types';
+import { type Location } from '@shopgate/engage/locations';
 import { address, addressIcon, title } from './CartItemGroup.style';
 
 type Props = {
