+++ libraries/engage/product/product.types.js
@@ -6,27 +6,3 @@ export type ProductId = string;
 export type Product = {
   id: ProductId;
 }
-
-export type AddToCartProduct = {
-  productId: ProductId;
-  quantity: number;
-  fulfillment?: {
-    method: string;
-    location: {
-      code: string;
-      name: string;
-    }
-  }
-}
-
-export type CartItem = {
-  cartItemId: string,
-  quantity: number,
-  fulfillment?: {
-    method: string;
-    location: {
-      code: string;
-      name: string;
-    }
-  }
-}
